+++ R/functions_utility.R
@@ -201,8 +201,8 @@ ReducePreprocess <- function(
     is.null(effectPos) || !is.null(effectPos) && max(effectPos) <= nEffects
   )
 
-  if (type == "withTime") {
-    addTime <- Map(
+  ReduceEffUpdates <- function(statsChange, eventTime) {
+    reduce <- Map(
       function(x, y) {
         lapply(
           x,
@@ -211,12 +211,14 @@ ReducePreprocess <- function(
               return(NULL)
             } # no changes, no problem
             if (nrow(z) == 1) {
-              return(cbind(time = y, z))
+              return(
+                if (type == "withTime") cbind(time = y, z) else z
+              )
             } # just one update, no problem
-
+            
             discard <- duplicated(z[, c("node1", "node2")], fromLast = TRUE)
             changes <- cbind(
-              time = rep(y, sum(!discard)),
+              time = if (type == "withTime") rep(y, sum(!discard)) else NULL,
               z[!discard, , drop = FALSE]
             )
             if (nrow(changes) == 1) {
@@ -227,58 +229,76 @@ ReducePreprocess <- function(
           }
         ) # multiple updates might be repeated, keep the last
       },
-      preproData$dependentStatsChange, preproData$eventTime
+      statsChange, eventTime
     )
-
-    outDependetStatChange <- lapply(
+    
+    return(lapply(
       seq_len(nEffects),
       function(i) {
-        Reduce(rbind, lapply(addTime, "[[", i))
+        Reduce(rbind, lapply(reduce, "[[", i))
       }
-    )
-  } else if (type == "withoutTime") {
-    whTime <- lapply(
-      preproData$dependentStatsChange,
-      function(x) {
-        lapply(
-          x,
-          function(z) {
-            if (is.null(z)) {
-              return(NULL)
-            } # no changes, no problem
-            if (nrow(z) == 1) {
-              return(z)
-            } # just one update, no problem
-
-            discard <- duplicated(z[, c("node1", "node2")], fromLast = TRUE)
-            changes <- z[!discard, , drop = FALSE]
-            if (nrow(changes) == 1) {
-              return(changes)
-            }
-            changes <- changes[order(changes[, "node1"], changes[, "node2"]), ]
-          }
-        )
-      }
-    ) # multiple updates might be repeated, keep the last
-
-
-    outDependetStatChange <- lapply(
-      seq_len(nEffects),
-      function(i) {
-        Reduce(rbind, lapply(whTime, "[[", i))
-      }
-    )
+    ))
   }
-
-  if (!is.null(effectPos)) {
-    return(outDependetStatChange[effectPos])
+  
+  outDependentStatChange <- ReduceEffUpdates(
+    preproData$dependentStatsChange,
+    preproData$eventTime[preproData$orderEvents == 1]
+  )
+  
+  if ((preproData$subModel == "rate" || preproData$model == "REM") &&
+      length(preproData$rightCensoredStatsChange) > 0) {
+    rightCensoredStatChange <- ReduceEffUpdates(
+      preproData$rightCensoredStatsChange,
+      preproData$eventTime[preproData$orderEvents == 2]
+    )
+    
+    # combine lists
+    reducedPrepro <- list()
+    for (ii in seq.int(length(outDependentStatChange))) {
+      reducedPrepro[[ii]] <- list(
+        dependent = outDependentStatChange[[ii]],
+        rightCensored = rightCensoredStatChange[[ii]]
+      )
+    }
+    
+    if (!is.null(effectPos)) {
+      return(reducedPrepro[effectPos])
+    } else {
+      return(reducedPrepro)
+    }
+  } else if (!is.null(effectPos)) {
+    return(outDependentStatChange[effectPos])
   } else {
-    return(outDependetStatChange)
+    return(outDependentStatChange)
   }
 }
 
-
-
+#' Expand a set of changes
+#' 
+#' given a `node` and a `replace` value, set the change to all the nodes in
+#' the nodes `set`. Add the `time` to the array if provided.
+#'
+#' @param nodes a numeric vector with the sanitize position of the nodes
+#' @param replace a numeric vector with the replace value
+#' @param time a numeric vector with the time-stamp when the changes happen
+#' @param set a numeric vector with the index id of the node set
+#' @param isTwoMode logical, whether self ties are allow or not 
+#'
+#' @return an array with columns `node1`, `node2`, `replace` and `time` 
+#' @noRd
+#'
+#' @examples
+#' fillChanges(c(1, 3), c(4, 8), NULL, 1:5)
+fillChanges <- function(nodes, replace, time, set, isTwoMode = FALSE) {
+  times <- ifelse(isTwoMode, length(set), length(set) - 1)
+  
+  cbind(
+    time = if (!is.null(time)) rep(time, each = times) else NULL,   
+    node1 = rep(nodes, each = times),
+    node2 = Reduce(c, lapply(nodes, \(x) set[!set %in% x])),
+    replace = rep(replace, each = times)
+  )
+}
 
 #' update a network (adjacency matrix)
 #'