+++ tests/testthat/test-effects_preprocessing_DyNAM_choice.R
@@ -1,6 +1,5 @@
-# # inertia/tie weighted ----
 test_that(
-  "inertia/tie compute correct preprocessing objects weighted",
+  "inertia/tie weighted preprocessing",
   {
     preproData <- estimate(
       depNetwork ~ inertia(networkState, weighted = TRUE) +
@@ -8,28 +7,9 @@ test_that(
       model = "DyNAM", subModel = "choice", # modelType = "DyNAM-M"
       preprocessingOnly = TRUE
     )
-    outDependetStatChange <- Reduce(rbind, mapply(
-      function(x, y, effectPos) {
-        if (is.null(x[[effectPos]])) {
-          return(NULL)
-        } # no changes, no problem
-        if (nrow(x[[effectPos]]) == 1) {
-          return(cbind(time = y, x[[effectPos]]))
-        } # just one update, no problem
-
-        discard <- duplicated(x[[effectPos]][, c("node1", "node2")], fromLast = TRUE)
-        changes <- cbind(time = y[1], x[[effectPos]][!discard, , drop = FALSE])
-        if (nrow(changes) == 1) {
-          return(changes)
-        }
-        # print(changes)
-        changes <- changes[order(changes[, "node1"], changes[, "node2"]), ]
-        return(changes) # multiple updates might be repeated, keep the last
-      },
-      preproData$dependentStatsChange, preproData$eventTime,
-      effectPos = 2
-    ))
-    expect_equal(preproData$initialStats[, , 1],
+    outDependentStatChange <- ReducePreprocess(preproData)
+    expect_equal(
+      preproData$initialStats[, , 1],
       matrix(c(
         0, 3, 0, 0, 0,
         1, 0, 1, 1, 0,
@@ -39,7 +19,8 @@ test_that(
       ), 5, 5, TRUE),
       label = "initialization of the statistics matrix"
     )
-    expect_equal(preproData$initialStats[, , 2],
+    expect_equal(
+      preproData$initialStats[, , 2],
       matrix(c(
         0, 0, 0, 1, 0,
         0, 0, 0, 0, 0,
@@ -49,15 +30,17 @@ test_that(
       ), 5, 5, TRUE),
       label = "initialization of the statistics matrix"
     )
-    expect_equal(Reduce(rbind, lapply(preproData$dependentStatsChange, "[[", 1)),
+    expect_equal(
+      Reduce(rbind, lapply(preproData$dependentStatsChange, "[[", 1)),
       cbind(
-        node1 = c(1, 3, 2, 2, 5, 1, 3, 3, 4, 2, 5),
-        node2 = c(2, 2, 3, 3, 1, 5, 4, 4, 2, 3, 2),
+        node1   = c(1, 3, 2, 2, 5, 1, 3, 3, 4, 2, 5),
+        node2   = c(2, 2, 3, 3, 1, 5, 4, 4, 2, 3, 2),
         replace = c(4, 2, 2, 3, 1, 2, 2, 3, 1, 4, 1)
       ),
       label = "updating with increment works"
     ) # n-1 updates
-    expect_equal(outDependetStatChange,
+    expect_equal(
+      outDependentStatChange[[2]],
       cbind(
         time = c(9, 15, 16, 19, 19, 28, 28),
         node1 = c(4, 2, 5, 4, 4, 1, 3),
@@ -66,19 +49,64 @@ test_that(
       ),
       label = "updating with increment works"
     ) # n-1 updates
+    expect_equal(
+      preproData$rightCensoredStatsChange,
+      list(),
+      label = "updating with increment works"
+    )
+    expect_equal(
+      preproData$intervals,
+      numeric(),
+      label = "intervals dependent"
+    ) # intervals are computed with right censored events
+    expect_equal(
+      preproData$rightCensoredIntervals,
+      numeric(),
+      label = "intervals right censored"
+    )
+    expect_equal(
+      preproData$orderEvents,
+      rep(1, nrow(eventsIncrement)),
+      label = "order events"
+    )
+    expect_equal(
+      preproData$eventTime,
+      eventsIncrement$time |> unique(),
+      label = "events times"
+    )
+    expect_equal(
+      preproData$eventSender,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrement$sender)),
+      label = "sender events"
+    )
+    expect_equal(
+      preproData$eventReceiver,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrement$receiver)),
+      label = "receiver events"
+    )
+    expect_equal(
+      preproData$startTime,
+      head(eventsIncrement$time, 1),
+      label = "start time"
+    )
+    expect_equal(
+      preproData$endTime,
+      tail(eventsIncrement$time, 1),
+      label = "end Time"
+    )
   }
 )
 
-# inertia not weighted ----
 test_that(
-  "inertia compute correct preprocessing objects no weighted",
+  "inertia  not weighted preprocessing",
   {
     preproData <- estimate(
       depNetwork ~ inertia,
       model = "DyNAM", subModel = "choice",
       preprocessingOnly = TRUE
     )
-    expect_equal(preproData$initialStats[, , 1],
+    expect_equal(
+      preproData$initialStats[, , 1],
       matrix(c(
         0, 1, 0, 0, 0,
         1, 0, 1, 1, 0,
@@ -88,7 +116,8 @@ test_that(
       ), 5, 5, TRUE),
       label = "initialization of the statistics matrix"
     )
-    expect_equal(Reduce(rbind, lapply(preproData$dependentStatsChange, "[[", 1)),
+    expect_equal(
+      Reduce(rbind, lapply(preproData$dependentStatsChange, "[[", 1)),
       cbind(
         node1 = c(3, 5, 1, 4, 5),
         node2 = c(2, 1, 5, 2, 2),
@@ -96,47 +125,70 @@ test_that(
       ),
       label = "updating with increment works"
     ) # n-1 updates
+    expect_equal(
+      preproData$rightCensoredStatsChange,
+      list(),
+      label = "updating with increment works"
+    )
+    expect_equal(
+      preproData$intervals,
+      numeric(),
+      label = "intervals dependent"
+    ) # intervals are computed with right censored events
+    expect_equal(
+      preproData$rightCensoredIntervals,
+      numeric(),
+      label = "intervals right censored"
+    )
+    expect_equal(
+      preproData$orderEvents,
+      rep(1, nrow(eventsIncrement)),
+      label = "order events"
+    )
+    expect_equal(
+      preproData$eventTime,
+      eventsIncrement$time |> unique(),
+      label = "events times"
+    )
+    expect_equal(
+      preproData$eventSender,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrement$sender)),
+      label = "sender events"
+    )
+    expect_equal(
+      preproData$eventReceiver,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrement$receiver)),
+      label = "receiver events"
+    )
+    expect_equal(
+      preproData$startTime,
+      head(eventsIncrement$time, 1),
+      label = "start time"
+    )
+    expect_equal(
+      preproData$endTime,
+      tail(eventsIncrement$time, 1),
+      label = "end Time"
+    )
   }
 )
 
-# # inertia windowed size 2 weighted ----
 test_that(
-  "inertia compute correct preprocessing objects windowed and weighted",
+  "inertia windowed and weighted preprocessing",
   {
-    skip_on_ci()
-    skip_on_cran()
-    skip_on_covr()
-    skip_on_bioc()
-    preproData <- estimate(depNetwork ~ inertia(networkState, weighted = TRUE, window = 2),
+    preproData <- estimate(
+      depNetwork ~ inertia(networkState, weighted = TRUE, window = 2),
       model = "DyNAM", subModel = "choice",
       preprocessingOnly = TRUE
     )
 
-    outDependetStatChange <- Reduce(rbind, mapply(
-      function(x, y) {
-        if (is.null(x[[1]])) {
-          return(NULL)
-        } # no changes, no problem
-        if (nrow(x[[1]]) == 1) {
-          return(cbind(time = y, x[[1]]))
-        } # just one update, no problem
-
-        discard <- duplicated(x[[1]][, c("node1", "node2")], fromLast = TRUE)
-        changes <- cbind(time = y[1], x[[1]][!discard, , drop = FALSE])
-        if (nrow(changes) == 1) {
-          return(changes)
-        }
-        # print(changes)
-        changes <- changes[order(changes[, "node1"], changes[, "node2"]), ]
-        return(changes) # multiple updates might be repeated, keep the last
-      },
-      preproData$dependentStatsChange, preproData$eventTime
-    ))
+    outDependentStatChange <- ReducePreprocess(preproData)[[1]]
     expect_equal(preproData$initialStats[, , 1],
       matrix(0, 5, 5, TRUE),
-      label = "initialization of the statistics matrix"
+      label = "initialization of the stats matrix"
     )
-    expect_equal(outDependetStatChange,
+    expect_equal(
+      outDependentStatChange,
       cbind(
         time = c(6, 9, 13, 15, 16, 16, 19, 19, 23, 28, 29, 32, 32, 36),
         node1 = c(1, 3, 2, 2, 2, 5, 1, 5, 3, 3, 4, 2, 4, 5),
@@ -145,5 +197,147 @@ test_that(
       ),
       label = "updating with increment works"
     ) # n-1 updates
+    expect_equal(
+      preproData$rightCensoredStatsChange,
+      list(),
+      label = "updating with increment works"
+    )
+    expect_equal(
+      preproData$intervals,
+      numeric(),
+      label = "intervals dependent"
+    ) # intervals are computed with right censored events
+    expect_equal(
+      preproData$rightCensoredIntervals,
+      numeric(),
+      label = "intervals right censored"
+    )
+    expect_equal(
+      preproData$orderEvents,
+      rep(1, nrow(eventsIncrement)),
+      label = "order events"
+    )
+    expect_equal(
+      preproData$eventTime,
+      eventsIncrement$time |> unique(),
+      label = "events times"
+    )
+    expect_equal(
+      preproData$eventSender,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrement$sender)),
+      label = "sender events"
+    )
+    expect_equal(
+      preproData$eventReceiver,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrement$receiver)),
+      label = "receiver events"
+    )
+    expect_equal(
+      preproData$startTime,
+      head(eventsIncrement$time, 1),
+      label = "start time"
+    )
+    expect_equal(
+      preproData$endTime,
+      tail(eventsIncrement$time, 1),
+      label = "end Time"
+    ) # end time includes non dependent events
+  }
+)
+
+test_that(
+  "inertia/tie startTime endTime preprocessing",
+  {
+    preproData <- estimate(
+      depNetwork ~ inertia(networkState, weighted = TRUE) +
+        tie(networkExog, weighted = TRUE),
+      model = "DyNAM", subModel = "choice", # modelType = "DyNAM-M"
+      preprocessingOnly = TRUE,
+      estimationInit = list(startTime = 10, endTime = 30)
+    )
+    outDependentStatChange <- ReducePreprocess(preproData)
+    eventsIncrementSubset <- subset(eventsIncrement, time >= 10 & time <= 30)
+    expect_equal(preproData$initialStats[, , 1],
+                 matrix(c(
+                   0, 4, 0, 0, 0,
+                   1, 0, 2, 1, 0,
+                   0, 2, 0, 1, 0,
+                   0, 0, 1, 0, 0,
+                   0, 0, 0, 0, 0
+                 ), 5, 5, TRUE),
+                 label = "initialization of the statistics matrix"
+    )
+    expect_equal(preproData$initialStats[, , 2],
+                 matrix(c(
+                   0, 0, 0, 1, 0,
+                   0, 0, 0, 0, 0,
+                   0, 2, 0, 0, 0,
+                   1, 1, 0, 0, 0,
+                   1, 2, 0, 0, 0
+                 ), 5, 5, TRUE),
+                 label = "initialization of the statistics matrix"
+    )
+    expect_equal(Reduce(rbind, lapply(preproData$dependentStatsChange, "[[", 1)),
+                 cbind(
+                   node1 =   c(2, 5, 1, 3, 3, 4),
+                   node2 =   c(3, 1, 5, 4, 4, 2),
+                   replace = c(3, 1, 2, 2, 3, 1)
+                 ),
+                 label = "updating with increment works"
+    ) # n-1 updates
+    expect_equal(outDependentStatChange[[2]],
+                 cbind(
+                   time = c(15, 16, 19, 19, 28, 28),
+                   node1 = c(2, 5, 4, 4, 1, 3),
+                   node2 = c(3, 1, 2, 5, 3, 5),
+                   replace = c(1, 4, 0, 1, 2, 3)
+                 ),
+                 label = "updating with increment works"
+    )
+    expect_equal(
+      preproData$rightCensoredStatsChange,
+      list(),
+      label = "updating with increment works"
+    )
+    expect_equal(
+      preproData$intervals,
+      numeric(),
+      label = "intervals dependent"
+    ) # intervals are computed with right censored events
+    expect_equal(
+      preproData$rightCensoredIntervals,
+      numeric(),
+      label = "intervals right censored"
+    )
+    expect_equal(
+      preproData$orderEvents,
+      rep(1, nrow(eventsIncrementSubset)),
+      label = "order events"
+    )
+    expect_equal(
+      preproData$eventTime,
+      eventsIncrementSubset$time |> unique(),
+      label = "events times"
+    )
+    expect_equal(
+      preproData$eventSender,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrementSubset$sender)),
+      label = "sender events"
+    )
+    expect_equal(
+      preproData$eventReceiver,
+      as.numeric(gsub("\\D+(\\d)", "\\1", eventsIncrementSubset$receiver)),
+      label = "receiver events"
+    )
+    expect_equal(
+      preproData$startTime,
+      10,
+      label = "start time"
+    )
+    expect_equal(
+      preproData$endTime,
+      30,
+      label = "end Time"
+    )
   }
 )